Authorization started to work, but can't create Professors anymore

+++ client/src/reducers/auth.js
@@ -1,19 +1,14 @@
-import {AUTH, LOGOUT} from '../constants/actionTypes.js'
-
 const authReducer = (state = {authData: null}, action) => {
-    switch (action.type) {
-        case AUTH:
-            localStorage.setItem('profile', JSON.stringify({...action?.data}));
-
-            return {...state, authData: action?.data};
-
-        case LOGOUT:
-            localStorage.clear();
-
-            return {...state, authData: null};
-        default:
-            return state;
+    if (action.type === "AUTH") {
+        localStorage.setItem('profile', JSON.stringify({...action?.data}));
+        return {...state, authData: action?.data};
+    } else if (action.type === "LOGOUT") {
+        localStorage.clear();
+        return {...state, authData: null};
+    } else {
+        return state;
     }
-};
+}
+
 
 export default authReducer;
